chore: code review & cleanup
