+++ tests/integration/test_token_metrics.py
@@ -714,7 +714,7 @@ class TestDialecticTokenMetrics:
         )
 
         with patch(
-            "src.dialectic.core.honcho_llm_call",
+            "src.dialectic.base.honcho_llm_call",
             new=AsyncMock(return_value=mock_response),
         ):
             await agent.answer("What do you know about this user?")
@@ -761,7 +761,7 @@ class TestDialecticTokenMetrics:
         )
 
         with patch(
-            "src.dialectic.core.honcho_llm_call",
+            "src.dialectic.base.honcho_llm_call",
             new=AsyncMock(return_value=mock_response),
         ):
             await agent.answer("What do you know about this user?")
@@ -824,7 +824,7 @@ class TestDialecticTokenMetrics:
         )
 
         with patch(
-            "src.dialectic.core.honcho_llm_call",
+            "src.dialectic.base.honcho_llm_call",
             new=AsyncMock(return_value=mock_response),
         ):
             await agent.answer("What do you know about this user?")
